glTF Test Module; glTF Refactor After Array 'len' attrib removal
Changes to the Array class required big refactor of the glTF class.
Tests all passing; updates to other container init() methods:
    (fixed bug where a nullptr to scratch allocator could be dereffed in
    StringBuffer and Array classes)

Add a small test module for glTF class to ensure refactor did not
introduce bugs. This module is not extensive, but tests a number of
different load functions, such as StringBuffer and Array loads, so it
works for now.